Fix animation speed multiplier, add research into animation layers
Speed multiplier = 0 freezes the animation. Problem is that this affects the animation on the next state transition. So be careful with speed multiplier.
Researching into using animation layers for Jill animation transition to make the acceleration from walking looks more gradual.
Researching into Vanguard  animation layers to lower the arms holding the gun.