Fix Null references
Shit hits the fan when I unpack the canvas and start assigning prefabs instead of actual instances to a script.  Unless the script belongs to the ancestor prefab node, and the descendant node gets assigned to the script, do not assign prefabs to a script of a totally unrelated node.
I also increased the priority of the M4Shotgun script so it will start before Vanguard. Vanguard needs the M4Shotgun script to start before its own start gets called.