Goal pointer
Add a goal pointer on the HUD. The goal pointer is oriented using the camera view. Unfortunately, for some weird reasons, the camera z axis produces a sprite jump when the goal is directly in front of it. My solution right now is to calculate the angle between the player and the camera, then subtract the angle between the player and the  goal  with this angle.  Using the player's forward axis produces a smoother goal pointing animation.